+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactoring custom header collection API for consistency
   ([#1064](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1064))
 
+### Added
+- `opentelemetry-instrument` and `opentelemetry-bootstrap` now include a `--version` flag
+  ([#1065](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1065))
+
 ## [1.11.1-0.30b1](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v1.11.1-0.30b1) - 2022-04-21
 
 ### Added